fix: make any motor go forward no matter if connected left or right (#58)
added a small inverted variable to each motor that will reverse the
state of it, if you set that motor forward it'll try to go backwards,
useful for when the motor is connected in the wrong side and it needs to
receive the opposite values. this will be later fixed in the actual
hardware.